Alpha 0.2 Update
This update took me quite literally, all day to program. As we swap from If-Else chains to Hashes, things get a bit more confusing. I ended up having to fix a stupid error with the 'assets' folder, and instead swapped to storing the noble gas and electron configurations inside of a array. Also, thanks to @MepiLeaf for helping me with some issues in my first hour!