Preparations for Alpha 0.3 and new commands
As we prepare for the next alpha release, I am starting to introduce more and more commands, so I made a small program that takes the function inside of the 'hash.c' file and lets you make hashes out of anything. You can compile that program from the 'build_win.bat' file.