Preparations for v0.2
Note: Compiling the program right now will return an error.

This update is focusing on putting all the commands into their own separate files, along with a API I'm going to be setting up so that you can easily put your own command in without having to hard code it in.

Bit confusing for now but I will get it to work soon.

Thanks,
- Jack