`Bunsen Code` | Working towards v0.2
Slowly getting there. I didn't like the way that the command folder worked, so instead I'm just going to be re-writing the cli.h file and putting it into the root directory.

I'm also going to be working on a new command handler system that might be more efficient.

Thanks,
- Jack